requirement.txt
allure-behave==2.8.16
allure-python-commons==2.8.16
Appium-Python-Client==0.44
appnope==0.1.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.7.1
behave==1.2.6
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
checkout==0.2.0
constant==0.0.4
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
grpcio==1.16.1
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.15
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.5
numpy==1.16.4
packaging==19.2
pandocfilters==1.4.2
parse==1.11.1
parse-type==0.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.13.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.3.1
pyparsing==2.4.5
pyrsistent==0.14.11
pytest==5.3.2
python-dateutil==2.8.0
python-dotenv==0.10.3
PyYAML==5.3
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.22.0
scipy==1.3.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
somewhere==0.0.1
soupsieve==1.9.1
terminado==0.8.1
testpath==0.4.2
torchfile==0.1.0
tornado==6.0.1
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==16.7.9
visdom==0.1.8.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
widgetsnbextension==3.4.2
wrapt==1.11.1
zipp==0.6.0
